refactor(view): unify events.
